added persistence helper and removed unused persistence methods
