--- shared/src/persistence/dynamo/cases/getCaseIrsPractitioners.js
@@ -1,26 +1,0 @@
-const { query } = require('../../dynamodbClientService');
-
-/**
- * getCaseIrsPractitioners
- *
- * @param {object} providers the providers object
- * @param {object} providers.applicationContext the application context
- * @returns {Function} async function to be used in an array.map
- */
-exports.getCaseIrsPractitioners = ({
-  applicationContext,
-}) => async theCase => ({
-  ...theCase,
-  irsPractitioners: await query({
-    ExpressionAttributeNames: {
-      '#pk': 'pk',
-      '#sk': 'sk',
-    },
-    ExpressionAttributeValues: {
-      ':pk': `case|${theCase.caseId}`,
-      ':prefix': 'irsPractitioner',
-    },
-    KeyConditionExpression: '#pk = :pk and begins_with(#sk, :prefix)',
-    applicationContext,
-  }),
-});
